refactor: move db into core, use yaml for types
