+++ sonar-core/lib/record.js
