+++ sonar-core/package.json
@@ -4,14 +4,14 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "tape test/*.js"
+    "test": "tape test/*.js",
+    "build": "protocol-buffers schema.proto > lib/messages.js"
   },
   "keywords": [],
   "author": "",
   "license": "GPL-3.0",
   "dependencies": {
     "@arso-project/sonar-tantivy": "^0.2.14",
-    "@arso-project/sonar-db": "^0.3.0-alpha.7",
     "corestore": "^5.6.0",
     "corestore-swarm-networking": "^5.5.1",
     "debug": "^4.1.1",
