+++ sonar-core/views/helpers.js
