+++ sonar-core/views/history.js
