+++ sonar-core/views/indexes.js
