+++ sonar-core/views/kv.js
