+++ sonar-core/views/records.js
