--- sonar-db/lib/messages.js
@@ -1,251 +1,0 @@
-// This file is auto generated by the protocol-buffers compiler
-
-/* eslint-disable quotes */
-/* eslint-disable indent */
-/* eslint-disable no-redeclare */
-/* eslint-disable camelcase */
-
-// Remember to `npm install --save protocol-buffers-encodings`
-var encodings = require('protocol-buffers-encodings')
-var varint = encodings.varint
-var skip = encodings.skip
-
-var Record = exports.Record = {
-  buffer: true,
-  encodingLength: null,
-  encode: null,
-  decode: null
-}
-
-var Header = exports.Header = {
-  buffer: true,
-  encodingLength: null,
-  encode: null,
-  decode: null
-}
-
-defineRecord()
-defineHeader()
-
-function defineRecord () {
-  Record.Op = {
-  PUT: 0,
-  DEL: 1
-}
-
-  var enc = [
-    encodings.string,
-    encodings.enum,
-    encodings.bytes,
-    encodings.varint
-  ]
-
-  Record.encodingLength = encodingLength
-  Record.encode = encode
-  Record.decode = decode
-
-  function encodingLength (obj) {
-    var length = 0
-    if (!defined(obj.id)) throw new Error("id is required")
-    var len = enc[0].encodingLength(obj.id)
-    length += 1 + len
-    if (!defined(obj.op)) throw new Error("op is required")
-    var len = enc[1].encodingLength(obj.op)
-    length += 1 + len
-    if (defined(obj.type)) {
-      var len = enc[0].encodingLength(obj.type)
-      length += 1 + len
-    }
-    if (defined(obj.value)) {
-      var len = enc[2].encodingLength(obj.value)
-      length += 1 + len
-    }
-    if (defined(obj.timestamp)) {
-      var len = enc[3].encodingLength(obj.timestamp)
-      length += 1 + len
-    }
-    if (defined(obj.links)) {
-      for (var i = 0; i < obj.links.length; i++) {
-        if (!defined(obj.links[i])) continue
-        var len = enc[0].encodingLength(obj.links[i])
-        length += 1 + len
-      }
-    }
-    return length
-  }
-
-  function encode (obj, buf, offset) {
-    if (!offset) offset = 0
-    if (!buf) buf = Buffer.allocUnsafe(encodingLength(obj))
-    var oldOffset = offset
-    if (!defined(obj.id)) throw new Error("id is required")
-    buf[offset++] = 10
-    enc[0].encode(obj.id, buf, offset)
-    offset += enc[0].encode.bytes
-    if (!defined(obj.op)) throw new Error("op is required")
-    buf[offset++] = 16
-    enc[1].encode(obj.op, buf, offset)
-    offset += enc[1].encode.bytes
-    if (defined(obj.type)) {
-      buf[offset++] = 26
-      enc[0].encode(obj.type, buf, offset)
-      offset += enc[0].encode.bytes
-    }
-    if (defined(obj.value)) {
-      buf[offset++] = 34
-      enc[2].encode(obj.value, buf, offset)
-      offset += enc[2].encode.bytes
-    }
-    if (defined(obj.timestamp)) {
-      buf[offset++] = 40
-      enc[3].encode(obj.timestamp, buf, offset)
-      offset += enc[3].encode.bytes
-    }
-    if (defined(obj.links)) {
-      for (var i = 0; i < obj.links.length; i++) {
-        if (!defined(obj.links[i])) continue
-        buf[offset++] = 50
-        enc[0].encode(obj.links[i], buf, offset)
-        offset += enc[0].encode.bytes
-      }
-    }
-    encode.bytes = offset - oldOffset
-    return buf
-  }
-
-  function decode (buf, offset, end) {
-    if (!offset) offset = 0
-    if (!end) end = buf.length
-    if (!(end <= buf.length && offset <= buf.length)) throw new Error("Decoded message is not valid")
-    var oldOffset = offset
-    var obj = {
-      id: "",
-      op: 0,
-      type: "",
-      value: null,
-      timestamp: 0,
-      links: []
-    }
-    var found0 = false
-    var found1 = false
-    while (true) {
-      if (end <= offset) {
-        if (!found0 || !found1) throw new Error("Decoded message is not valid")
-        decode.bytes = offset - oldOffset
-        return obj
-      }
-      var prefix = varint.decode(buf, offset)
-      offset += varint.decode.bytes
-      var tag = prefix >> 3
-      switch (tag) {
-        case 1:
-        obj.id = enc[0].decode(buf, offset)
-        offset += enc[0].decode.bytes
-        found0 = true
-        break
-        case 2:
-        obj.op = enc[1].decode(buf, offset)
-        offset += enc[1].decode.bytes
-        found1 = true
-        break
-        case 3:
-        obj.type = enc[0].decode(buf, offset)
-        offset += enc[0].decode.bytes
-        break
-        case 4:
-        obj.value = enc[2].decode(buf, offset)
-        offset += enc[2].decode.bytes
-        break
-        case 5:
-        obj.timestamp = enc[3].decode(buf, offset)
-        offset += enc[3].decode.bytes
-        break
-        case 6:
-        obj.links.push(enc[0].decode(buf, offset))
-        offset += enc[0].decode.bytes
-        break
-        default:
-        offset = skip(prefix & 7, buf, offset)
-      }
-    }
-  }
-}
-
-function defineHeader () {
-  var enc = [
-    encodings.string,
-    encodings.bytes
-  ]
-
-  Header.encodingLength = encodingLength
-  Header.encode = encode
-  Header.decode = decode
-
-  function encodingLength (obj) {
-    var length = 0
-    if (!defined(obj.type)) throw new Error("type is required")
-    var len = enc[0].encodingLength(obj.type)
-    length += 1 + len
-    if (defined(obj.metadata)) {
-      var len = enc[1].encodingLength(obj.metadata)
-      length += 1 + len
-    }
-    return length
-  }
-
-  function encode (obj, buf, offset) {
-    if (!offset) offset = 0
-    if (!buf) buf = Buffer.allocUnsafe(encodingLength(obj))
-    var oldOffset = offset
-    if (!defined(obj.type)) throw new Error("type is required")
-    buf[offset++] = 10
-    enc[0].encode(obj.type, buf, offset)
-    offset += enc[0].encode.bytes
-    if (defined(obj.metadata)) {
-      buf[offset++] = 18
-      enc[1].encode(obj.metadata, buf, offset)
-      offset += enc[1].encode.bytes
-    }
-    encode.bytes = offset - oldOffset
-    return buf
-  }
-
-  function decode (buf, offset, end) {
-    if (!offset) offset = 0
-    if (!end) end = buf.length
-    if (!(end <= buf.length && offset <= buf.length)) throw new Error("Decoded message is not valid")
-    var oldOffset = offset
-    var obj = {
-      type: "",
-      metadata: null
-    }
-    var found0 = false
-    while (true) {
-      if (end <= offset) {
-        if (!found0) throw new Error("Decoded message is not valid")
-        decode.bytes = offset - oldOffset
-        return obj
-      }
-      var prefix = varint.decode(buf, offset)
-      offset += varint.decode.bytes
-      var tag = prefix >> 3
-      switch (tag) {
-        case 1:
-        obj.type = enc[0].decode(buf, offset)
-        offset += enc[0].decode.bytes
-        found0 = true
-        break
-        case 2:
-        obj.metadata = enc[1].decode(buf, offset)
-        offset += enc[1].decode.bytes
-        break
-        default:
-        offset = skip(prefix & 7, buf, offset)
-      }
-    }
-  }
-}
-
-function defined (val) {
-  return val !== null && val !== undefined && (typeof val !== 'number' || !isNaN(val))
-}
