--- sonar-db/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@arso-project/sonar-db",
-  "version": "0.3.0-alpha.7",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test/*.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "GPL-3.0",
-  "dependencies": {
-    "ajv": "^6.12.2",
-    "base32": "^0.0.6",
-    "charwise": "^3.0.1",
-    "debug": "^4.1.1",
-    "level-live": "github:Frando/level-live#key-encodings",
-    "level-mem": "^5.0.1",
-    "pretty-hash": "^1.0.1",
-    "protocol-buffers-encodings": "^1.1.0",
-    "subleveldown": "^5.0.0",
-    "through2": "^3.0.1",
-    "thunky": "^1.1.0",
-    "unordered-materialized-kv": "^1.3.0"
-  },
-  "devDependencies": {
-    "tape": "^5.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
